refactor: extract parameter construction logic into a separate function for clarity and reusability
