+++ src/market_data_ingestor/src/providers/alpaca_rest/provider.rs
@@ -3,7 +3,7 @@ use reqwest::{header, Client};
 use secrecy::{ExposeSecret, SecretString};
 use shared_utils::env::get_env_var;
 
-use crate::{models::{bar::Bar, bar_series::BarSeries, request_params::{BarsRequestParams, ProviderParams}, timeframe::{TimeFrame, TimeFrameUnit}}, providers::{alpaca_rest::{params::AlpacaBarsParams, response::AlpacaResponse}, DataProvider, ProviderError, ProviderInitError}};
+use crate::{models::{bar::Bar, bar_series::BarSeries, request_params::{BarsRequestParams, ProviderParams}, timeframe::{TimeFrame, TimeFrameUnit}}, providers::{alpaca_rest::{params::{construct_params, AlpacaBarsParams}, response::AlpacaResponse}, DataProvider, ProviderError, ProviderInitError}};
 
 
 
@@ -44,56 +44,13 @@ impl AlpacaProvider {
     }
 }
 
-fn format_timeframe(tf: &TimeFrame) -> String {
-    let unit_str = match tf.unit {
-        TimeFrameUnit::Minute => "Min",
-        TimeFrameUnit::Hour => "Hour",
-        TimeFrameUnit::Day => "Day",
-        TimeFrameUnit::Week => "Week",
-        TimeFrameUnit::Month => "Month",
-    };
-    format!("{}{}", tf.amount, unit_str)
-}
+
 
 #[async_trait]
 impl DataProvider for AlpacaProvider {
     async fn fetch_bars(&self, params: BarsRequestParams) -> Result<Vec<BarSeries>, ProviderError>{
         // TODO: Implement Alpaca-specific timeframe validation here.
-
-        let symbols = params.symbols.join(",");
-        let timeframe = format_timeframe(&params.timeframe);
-
-        // TODO: Implement pagination using `next_page_token`.
-        // This initial implementation only fetches the first page.
-
-        let alpaca_params = match &params.provider_specific {
-            ProviderParams::Alpaca(p) => p.clone(),
-            _ => AlpacaBarsParams::default(),
-        };
-
-        let mut query_params = vec![
-            ("symbols".to_string(), symbols),
-            ("timeframe".to_string(), timeframe),
-            ("start".to_string(), params.start.to_rfc3339()),
-            ("end".to_string(), params.end.to_rfc3339()),
-        ];
-
-        // Add optional parameters if they exist
-        if let Some(adjustment) = alpaca_params.adjustment {
-            query_params.push(("adjustment".to_string(), serde_json::to_string(&adjustment).unwrap().replace('"', "")));
-        }
-        if let Some(feed) = alpaca_params.feed {
-            query_params.push(("feed".to_string(), serde_json::to_string(&feed).unwrap().replace('"', "")));
-        }
-        if let Some(currency) = alpaca_params.currency {
-            query_params.push(("currency".to_string(), currency));
-        }
-        if let Some(limit) = alpaca_params.limit {
-            query_params.push(("limit".to_string(), limit.to_string()));
-        }
-        if let Some(sort) = alpaca_params.sort {
-            query_params.push(("sort".to_string(), serde_json::to_string(&sort).unwrap().replace('"', "")));
-        }
+        let query_params = construct_params(&params);
 
         let response = self
             .client
